Error soolucionado

+++ fun.h
@@ -5,7 +5,7 @@
 
 extern double gendist (float *vec1, float *vec2);
 extern void grupo_cercano (int nvec, float mvec[][NDIM], float cent[][NDIM], int *popul);
-extern double silhouette_simple(float mvec[][NDIM], struct lista_grupos *listag, float cent[][NDIM], double a[]);
+extern double silhouette_simple(float mvec[][NDIM], struct lista_grupos *listag, float cent[][NDIM], float a[]);
 extern void analisis_campos(struct lista_grupos *listag, float mcam[][NCAM], struct analisis *info_cam);
 
 extern void inicializar_centroides(float cent[][NDIM]);
